test/plugins/test_linux_log_parser: Fix unit test test_metadata_creation
Add the timestamp `[ 0.0 ]` so the log for the test_metadata_creation
unit test will be matched by the log parser.

Previously the log would be thrown out by the `__kernel_msgs_only`
method of the log parser plugin, which removes any lines which do not
contain a timestamp. This test would pass without the timestamp since
the log parser currently creates metadata even when there are no log
parser matches. However, in future we may wish to only create tests when
there are log parser matches, which would cause this unit test to
perform incorrectly.

Signed-off-by: Katie Worton <katie.worton@linaro.org>