plugins/lib/base_log_parser: Update truncating of test name
Update the code that truncated the test name to match the max test name
length in SQUAD. Originally this was in the "create_name" method but it
has now been moved to "create_name_log_dict" as this is the place where
the final SQUAD name is built. Additionally, this patch adds support for
SHAs being optional, allowing for a longer name if SHAs are not being
used.

Signed-off-by: Katie Worton <katie.worton@linaro.org>