ci/models.py: silence duplicated job warning
When SQUAD attempts to fetch a job that already exist in a build, it raises an error,
thus triggering an event in Sentry. When lots of this events happen, lots of Sentry
notifications do too, overwhelming engineers.

This patch records the error in the job.failure field and avoid logging it as error.

Signed-off-by: Charles Oliveira <charles.oliveira@linaro.org>