Dockerfile: bump Python version to 3.11
Python 3.11 is the stadard one on Debian bookworm.

BIG NOTE: In recent versions of Python, the pip package enforces that packages are installed
within a virtual environment (venv). While reading on it, that seems to be a good practice nowadays.

For that matter, it was either maintaing non-debian python packages in debian or install them via pip.
While having debian-like packages is great for stability, maintenance wouldn't be so great. The downside
with installing dependencies via pip packages is that sometimes we might get unexpected updates and things
could break. Still, CI/CD should be able to catch those.

Signed-off-by: Charles Oliveira <charles.oliveira@linaro.org>