ci: backend: tuxsuite: download attachments
When sending email reports, LKFT has been pointing to config and
reproducers that is stored in TuxSuite. However, the retention policy
for TuxSuite is set to 30 days. If config and the reproducer is stored
in SQUAD the artefacts can be re-created easily.
Without the config and reproducers its harder to recreate teh artefacts.

Signed-off-by: Anders Roxell <anders.roxell@linaro.org>