ci: tuxsuite: don't ignore the lava "suite"
The "lava" suite contains results created from the different actions in
the LAVA job. While failures in those are usually not something we need
to be reported upstream, triaging engineers *need* to be aware of
potential problems in the LAVA jobs.

A contradiction is that we are now ignoring those in our own test suite.
:-)

Signed-off-by: Antonio Terceiro <antonio.terceiro@linaro.org>