ci/backend/tuxsuite.py: improve memory consumption
When using `yaml.load_safe()` on large YAML files might result in huge memory
consumption. For instance, we had a SQUAD instance parsing a LAVA YAML log file
of 24MB and that took 1.3GB and ~10 minutes just to parse into Python objects.

The reasoning behind is because Python adds extra properties when deserializing
yaml objects into memory.

Since we know that each line in the YAML log is a line in the log, it is fairly
trivial to retrieve what we need and ignore the rest, avoiding the whole YAML
parsing.

Signed-off-by: Charles Oliveira <charles.oliveira@linaro.org>